# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jikan API",
...
        "termsOfService": "https://jikan.moe/terms",
        "contact": {
            "name": "API Support (Discord)",
            "url": "http://discord.jikan.moe"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/jikan-me/jikan-rest/blob/master/LICENSE"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "https://api.jikan.moe/v4",
            "description": "Jikan REST API Beta"
        }
    ],
    "paths": {
...
        "/anime": {
            "get": {
                "tags": [
                    "anime"
                ],
                "operationId": "getAnimeSearch",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/limit"
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/anime_search_query_type"
                        }
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "min_score",
                        "in": "query",
                        "description": "Set a minimum score for results.",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_score",
                        "in": "query",
                        "description": "Set a maximum score for results",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/anime_search_query_status"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/anime_search_query_rating"
                        }
                    },
                    {
                        "name": "sfw",
                        "in": "query",
                        "description": "Filter out Adult entries",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "genres",
                        "in": "query",
                        "description": "Filter by genre(s) IDs. Can pass multiple with a comma as a delimiter. e.g 1,2,3",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "genres_exclude",
                        "in": "query",
                        "description": "Exclude genre(s) IDs. Can pass multiple with a comma as a delimiter. e.g 1,2,3",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/anime_search_query_orderby"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/search_query_sort"
                        }
                    },
                    {
                        "name": "letter",
                        "in": "query",
                        "description": "Return entries starting with the given letter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "producers",
                        "in": "query",
                        "description": "Filter by producer(s) IDs. Can pass multiple with a comma as a delimiter. e.g 1,2,3",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns search results for anime",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/anime_search"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/manga": {
            "get": {
                "tags": [
                    "manga"
                ],
                "operationId": "getMangaSearch",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/limit"
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/manga_search_query_type"
                        }
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "min_score",
                        "in": "query",
                        "description": "Set a minimum score for results.",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_score",
                        "in": "query",
                        "description": "Set a maximum score for results",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/manga_search_query_status"
                        }
                    },
                    {
                        "name": "sfw",
                        "in": "query",
                        "description": "Filter out Adult entries",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "genres",
                        "in": "query",
                        "description": "Filter by genre(s) IDs. Can pass multiple with a comma as a delimiter. e.g 1,2,3",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "genres_exclude",
                        "in": "query",
                        "description": "Exclude genre(s) IDs. Can pass multiple with a comma as a delimiter. e.g 1,2,3",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/manga_search_query_orderby"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/search_query_sort"
                        }
                    },
                    {
                        "name": "letter",
                        "in": "query",
                        "description": "Return entries starting with the given letter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "magazines",
                        "in": "query",
                        "description": "Filter by magazine(s) IDs. Can pass multiple with a comma as a delimiter. e.g 1,2,3",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns search results for manga",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/manga_search"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
...
        "/users/userbyid": {
            "get": {
                "tags": [
                    "users"
                ],
                "operationId": "getUserById",
                "responses": {
                    "200": {
                        "description": "Returns username by ID search",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/user_by_id"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "search_query_sort": {
                "description": "Characters Search Query Sort",
                "type": "string",
                "enum": [
                    "desc",
                    "asc"
                ]
            },
...
            "anime_search_query_type": {
                "description": "Available Anime types",
                "type": "string",
                "enum": [
                    "tv",
                    "movie",
                    "ova",
                    "special",
                    "ona",
                    "music"
                ]
            },
            "anime_search_query_status": {
                "description": "Available Anime statuses",
                "type": "string",
                "enum": [
                    "airing",
                    "complete",
                    "upcoming"
                ]
            },
            "anime_search_query_rating": {
                "description": "Available Anime audience ratings<br><br><b>Ratings</b><br><ul><li>G - All Ages</li><li>PG - Children</li><li>PG-13 - Teens 13 or older</li><li>R - 17+ (violence & profanity)</li><li>R+ - Mild Nudity</li><li>Rx - Hentai</li></ul>",
                "type": "string",
                "enum": [
                    "g",
                    "pg",
                    "pg13",
                    "r17",
                    "r",
                    "rx"
                ]
            },
...
        },
...
        "parameters": {
            "page": {
                "name": "page",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
            "limit": {
                "name": "limit",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            }
        }
    },
...
}